requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
requests==2.31.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
# Install audiocraft after torch
audiocraft @ git+https://github.com/facebookresearch/audiocraft#egg=audiocraft
